+++ cookbook/recipes/seismic_epic2d_iterate.py
@@ -34,8 +34,8 @@ if len(src) > 1:
     
 log.info("Generating synthetic travel-time data")
 srcs, recs = ft.utils.connect_points(src, rec_points)
-ptime = ft.seis.traveltime.straight_ray_2d(model, 'vp', srcs, recs)
-stime = ft.seis.traveltime.straight_ray_2d(model, 'vs', srcs, recs)
+ptime = ft.seis.ttime2d.straight(model, 'vp', srcs, recs)
+stime = ft.seis.ttime2d.straight(model, 'vs', srcs, recs)
 error_level = 0.1
 ttr_true = stime - ptime
 ttr, error = ft.utils.contaminate(ttr_true, error_level, percent=True,